# s3_source: Lazy load fog library