--- a/lib/chef/knife/s3_source.rb
+++ b/lib/chef/knife/s3_source.rb
@@ -1,5 +1,3 @@
-require 'fog'
-
 class Chef
   class Knife
     class S3Source
@@ -30,6 +28,7 @@ class Chef
       end
 
       def fog
+        require 'fog' # lazy load the fog library to speed up the knife run
         @fog ||= Fog::Storage::AWS.new(
           aws_access_key_id: Chef::Config[:knife][:aws_access_key_id],
           aws_secret_access_key: Chef::Config[:knife][:aws_secret_access_key]