This speeds up all knife commands on systems where knife-ec2 plugin is
installed.

Benchmarks ( done with 'time' command):

| Command | Before | After | % Improvement |
|-------|-----|----|----|
knife help | 8.523s | 2.616s  | 69% |
knife search "chef_environment:prod" | 20.429s | 15.094 | 26% |
knife search "chef_environment:prod AND role:admin" | 10.625   | 5.125s | 51% |