@Igorshp 2015-05-13
As far as I can see, this does not break s3 functionality, but i'm struggling to identify a correct way to test it. Any suggestions?