@btm 2015-05-13
Looking at other code, I'd try something like this. I'm not sure it works, so I'd test it against master too and make sure it fails.

```
it "loads deps all lazily" do
  expect(Chef::Knife::S3Source).not_to receive(:require)
  Chef::Knife::S3Source.new
end
```