@Igorshp 2015-05-13
Thanks @btm, i like the idea, but unforunately it doesn't work. 
I believe it's because that checks for a method called 'require' on S3Source class and real 'require' is part of ruby kernel? https://github.com/ruby/ruby/blob/ca6b174078fa15f33655be704d9409fdbc4f9929/lib/rubygems/core_ext/kernel_require.rb#L38

However:
``` 
expect(Kernel).not_to receive(:require)
```
also doesn't seem to work