@Igorshp 2015-05-13
Correct me if I'm wrong, but I'm starting to think that this is impossible to test as s3_source.rb is loaded and parsed when rspec loads.
So by the time it gets to tests, top level 'require' blocks have already been processed and do not get evaluated again.

It is also not possible to check if Fog::* is defined as it is required at the top of the rspec test and is used later for mocking.