@btm 2015-05-13
Yeah, might be impossible. Maybe for the second idea you move the require into the describe block somewhere and create a second describe block? I poked around the tests in core for the knife lazy loader stuff and I didn't see anything magical.