@Igorshp 2015-05-13
I'll try that.

I've also revised the PR. after researching how S3Source is used and how load_deps work, the original pr wouldn't have worked as run method is never called for S3Source.

Moving require to 'fog' method seems like a much simpler solution.