@Igorshp 2015-05-19
@thom @btm 
I've added 2 tests, but there is an interesting caveat.

first test tries to run fog and expects name error to show that library doesn't get loaded by default
second test, runs S3Source.fog and expects argument error to show that :fog method loads fog

If you run s3_source_deps_spec.rb on its own, it works great.

But! if it gets run as part of global tests with 'rspec', the first test fails, as fog is loaded in another spec and gets carried over.

this dependency leak essentially invalidates both of the tests.