@btm 2015-05-19
@Igorshp Oh yeah. I think we're totally in territory where what you're doing is a great exercise, but not required.

I could see a couple options if you want to pursue tests, but I'd be fine adding a comment about the require saying that we don't want to load fog right off.

You could run the test in shell like `chef/spec/integration/client/client_spec.rb` does, where you call out to ruby and make sure it doesn't raise / exit with an error.

You could add an rspec filter so the test doesn't run if Fog is already loaded (which probably means the test almost never runs, but has some documentation value).