@Igorshp 2015-05-19
@btm makes sense, I think that's a good solution.
I've added a spec filter and a couple of comments, it works well.

Will squash the commits once everyone is happy with the PR.