@adamedx 2015-05-19
:+1: as long as tests all pass and we have coverage that the lazy load actually works. In general, anything we can do to speed up knife subcommands is welcome, particularly on Windows.