@btm 2015-05-19
@Igorshp :+1:  rebase and I'll merge.