@Igorshp 2015-05-19
@btm all done. squashed commits, updated the commit message